Update hmmer to 3.4

+++ assembler/ext/src/hmmer/src/hmmbuild.c
@@ -1,6 +1,6 @@
 /* Profile HMM construction from a multiple sequence alignment
  */
-#include "p7_config.h"
+#include <p7_config.h>
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -558,7 +558,7 @@ usual_master(const ESL_GETOPTS *go, struct cfg_s *cfg)
 #endif
 
   infocnt = (ncpus == 0) ? 1 : ncpus;
-  ESL_ALLOC(info, sizeof(*info) * infocnt);
+  ESL_ALLOC(info, (ptrdiff_t) sizeof(*info) * infocnt);
 
   for (i = 0; i < infocnt; ++i)
   {
