+++ assembler/ext/src/hmmer/src/impl_neon/p7_omx.c
@@ -6,13 +6,13 @@
  *
  * ML, Fri Mar 12 10:26:31 2021 [Heidelberg]
  */
-#include "p7_config.h"
+#include <p7_config.h>
 
 #include <stdio.h>
 #include <math.h>
 #include <float.h>
 
-#include <arm_neon.h>		/* NEON */
+#include <arm_neon.h>           /* NEON */
 
 #include "easel.h"
 #include "esl_alphabet.h"
@@ -75,9 +75,9 @@ p7_omx_Create(int allocM, int allocL, int allocXL)
   ox->allocQ4  = p7O_NQF(allocM);
   ox->allocQ8  = p7O_NQW(allocM);
   ox->allocQ16 = p7O_NQB(allocM);
-  ox->ncells   = ox->allocR * ox->allocQ4 * 4;      /* # of DP cells allocated, where 1 cell contains MDI */
+  ox->ncells   = (int64_t) ox->allocR * (int64_t) ox->allocQ4 * 4;      /* # of DP cells allocated, where 1 cell contains MDI */
 
-  ESL_ALLOC(ox->dp_mem, sizeof(uint8x16_t) * ox->allocR * ox->allocQ4 * p7X_NSCELLS + 15);  /* floats always dominate; +15 for alignment */
+  ESL_ALLOC(ox->dp_mem, sizeof(uint8x16_t) * (int64_t) ox->allocR * (int64_t) ox->allocQ4 * p7X_NSCELLS + 15);  /* floats always dominate; +15 for alignment */
   ESL_ALLOC(ox->dpb,    sizeof(uint8x16_t  *) * ox->allocR);
   ESL_ALLOC(ox->dpw,    sizeof(int16x8_t   *) * ox->allocR);
   ESL_ALLOC(ox->dpf,    sizeof(float32x4_t *) * ox->allocR);
@@ -87,9 +87,9 @@ p7_omx_Create(int allocM, int allocL, int allocXL)
   ox->dpf[0] = (float32x4_t *) ( ( (unsigned long int) ((char *) ox->dp_mem + 15) & (~0xf)));
 
   for (i = 1; i <= allocL; i++) {
-    ox->dpf[i] = ox->dpf[0] + i * ox->allocQ4  * p7X_NSCELLS;
-    ox->dpw[i] = ox->dpw[0] + i * ox->allocQ8  * p7X_NSCELLS;
-    ox->dpb[i] = ox->dpb[0] + i * ox->allocQ16;
+    ox->dpf[i] = ox->dpf[0] + (int64_t) i * (int64_t) ox->allocQ4  * p7X_NSCELLS;
+    ox->dpw[i] = ox->dpw[0] + (int64_t) i * (int64_t) ox->allocQ8  * p7X_NSCELLS;
+    ox->dpb[i] = ox->dpb[0] + (int64_t) i * (int64_t) ox->allocQ16;
   }
 
   ox->allocXR = allocXL+1;
@@ -99,7 +99,7 @@ p7_omx_Create(int allocM, int allocL, int allocXL)
   ox->M              = 0;
   ox->L              = 0;
   ox->totscale       = 0.0;
-  ox->has_own_scales = TRUE;	/* most matrices are Forward, control their own scale factors */
+  ox->has_own_scales = TRUE;    /* most matrices are Forward, control their own scale factors */
 #if eslDEBUGLEVEL > 0
   ox->debugging = FALSE;
   ox->dfp       = NULL;
@@ -133,14 +133,14 @@ p7_omx_Create(int allocM, int allocL, int allocXL)
 int
 p7_omx_GrowTo(P7_OMX *ox, int allocM, int allocL, int allocXL)
 {
-  void  *p;
-  int    nqf  = p7O_NQF(allocM);	       /* segment length; total # of striped vectors for uchar */
-  int    nqw  = p7O_NQW(allocM);	       /* segment length; total # of striped vectors for float */
-  int    nqb  = p7O_NQB(allocM);	       /* segment length; total # of striped vectors for float */
-  size_t ncells = (allocL+1) * nqf * 4;
-  int    reset_row_pointers = FALSE;
-  int    i;
-  int    status;
+  void   *p;
+  int     nqf    = p7O_NQF(allocM);            /* segment length; total # of striped vectors for uchar */
+  int     nqw    = p7O_NQW(allocM);            /* segment length; total # of striped vectors for float */
+  int     nqb    = p7O_NQB(allocM);            /* segment length; total # of striped vectors for float */
+  int64_t ncells = (int64_t) (allocL+1) * (int64_t) nqf * 4;
+  int     reset_row_pointers = FALSE;
+  int     i;
+  int     status;
  
   /* If all possible dimensions are already satisfied, the matrix is fine */
   if (ox->allocQ4*4 >= allocM && ox->validR > allocL && ox->allocXR >= allocXL+1) return eslOK;
@@ -150,7 +150,7 @@ p7_omx_GrowTo(P7_OMX *ox, int allocM, int allocL, int allocXL)
    */
   if (ncells > ox->ncells)
     {
-      ESL_RALLOC(ox->dp_mem, p, sizeof(uint8x16_t) * (allocL+1) * nqf * p7X_NSCELLS + 15);
+      ESL_RALLOC(ox->dp_mem, p, sizeof(uint8x16_t) * (int64_t) (allocL+1) * (int64_t) nqf * p7X_NSCELLS + 15);
       ox->ncells = ncells;
       reset_row_pointers = TRUE;
     }
@@ -193,9 +193,9 @@ p7_omx_GrowTo(P7_OMX *ox, int allocM, int allocL, int allocXL)
       ox->validR = ESL_MIN( ox->ncells / (nqf * 4), ox->allocR);
       for (i = 1; i < ox->validR; i++)
 	{
-	  ox->dpb[i] = ox->dpb[0] + i * nqb;
-	  ox->dpw[i] = ox->dpw[0] + i * nqw * p7X_NSCELLS;
-	  ox->dpf[i] = ox->dpf[0] + i * nqf * p7X_NSCELLS;
+	  ox->dpb[i] = ox->dpb[0] + (int64_t) i * (int64_t) nqb;
+	  ox->dpw[i] = ox->dpw[0] + (int64_t) i * (int64_t) nqw * p7X_NSCELLS;
+	  ox->dpf[i] = ox->dpf[0] + (int64_t) i * (int64_t) nqf * p7X_NSCELLS;
 	}
 
       ox->allocQ4  = nqf;
